fix: resolve linker errors and test failures in kernel_bpf CI
